move simulation clone logic to separate module
