+++ custom_components/span_panel/config_flow.py
@@ -73,9 +73,7 @@ from .options import (
     INVERTER_LEG2,
     POWER_DISPLAY_PRECISION,
 )
-from .simulation_generator import (
-    SimulationYamlGenerator,
-)  # lazy import to keep CF lean
+from .simulation_utils import clone_panel_to_simulation
 from .span_panel_api import SpanPanelApi
 
 _LOGGER = logging.getLogger(__name__)
@@ -1559,122 +1557,26 @@ class OptionsFlowHandler(config_entries.OptionsFlow):
         self, user_input: dict[str, Any] | None = None
     ) -> ConfigFlowResult:
         """Clone the live panel into a simulation YAML stored in simulation_configs."""
-        errors: dict[str, str] = {}
+        result = await clone_panel_to_simulation(self.hass, self.config_entry, user_input)
 
-        # Resolve coordinator (live)
-        coordinator_data = self.hass.data.get(DOMAIN, {}).get(self.config_entry.entry_id, {})
-        coordinator = coordinator_data.get(COORDINATOR)
-        if coordinator is None:
-            return self.async_abort(reason="coordinator_unavailable")
+        # If result is a ConfigFlowResult, return it directly
+        if hasattr(result, "type"):
+            return result  # type: ignore[return-value]
 
-        # Compute default filename
+        # Otherwise, result is (dest_path, errors) for the form
+        if isinstance(result, tuple) and len(result) == 2:
+            dest_path, errors = result
+            if not isinstance(errors, dict):
+                errors = {}
+        else:
+            # Fallback if result format is unexpected
+            _LOGGER.error(
+                "Unexpected result format from clone_panel_to_simulation: %s", type(result)
+            )
+            return self.async_abort(reason="unknown")
+
+        # Compute device name for form display
         device_name = self.config_entry.data.get("device_name", self.config_entry.title)
-        safe_device = slugify(device_name) if isinstance(device_name, str) else "span_panel"
-
-        # Default assumption when we can't infer (recomputed later from tabs)
-        num_tabs = 32
-        config_dir = Path(__file__).parent / "simulation_configs"
-        base_name = f"simulation_config_{safe_device}_{num_tabs}_circuit.yaml"
-        dest_path = config_dir / base_name
-
-        # Suffix if exists: _2, _3, ...
-        suffix_index = 1
-        if await self.hass.async_add_executor_job(dest_path.exists):
-            suffix_index = 2
-            while True:
-                candidate = (
-                    config_dir
-                    / f"simulation_config_{safe_device}_{num_tabs}_circuit_{suffix_index}.yaml"
-                )
-                if not await self.hass.async_add_executor_job(candidate.exists):
-                    dest_path = candidate
-                    break
-                suffix_index += 1
-
-        if user_input is not None:
-            try:
-                # Use a separate generator to build YAML purely from live data
-                # Pass solar leg selections from options if present
-                leg1_opt = self.config_entry.options.get(INVERTER_LEG1, 0)
-                leg2_opt = self.config_entry.options.get(INVERTER_LEG2, 0)
-                generator = SimulationYamlGenerator(
-                    hass=self.hass,
-                    coordinator=coordinator,
-                    solar_leg1=int(leg1_opt) if leg1_opt else None,
-                    solar_leg2=int(leg2_opt) if leg2_opt else None,
-                )
-                snapshot_yaml, num_tabs = await generator.build_yaml_from_live_panel()
-
-                def infer_template_for(name: str, tabs: list[int]) -> str:
-                    lname = str(name).lower()
-                    if any(k in lname for k in ["light", "lights"]):
-                        return "lighting"
-                    if "kitchen" in lname and "outlet" in lname:
-                        return "kitchen_outlets"
-                    if any(
-                        k in lname
-                        for k in ["hvac", "furnace", "air conditioner", "ac", "heat pump"]
-                    ):
-                        return "hvac"
-                    if any(k in lname for k in ["fridge", "refrigerator", "wine fridge"]):
-                        return "refrigerator"
-                    if any(k in lname for k in ["ev", "charger"]):
-                        return "ev_charger"
-                    if any(k in lname for k in ["pool", "spa", "fountain"]):
-                        return "pool_equipment"
-                    if any(k in lname for k in ["internet", "router", "network", "modem"]):
-                        return "always_on"
-                    # Heuristics: 240V multi-tab loads as major appliances
-                    if len(tabs) >= 2:
-                        return "major_appliance"
-                    # Fallbacks
-                    if "outlet" in lname:
-                        return "outlets"
-                    return "major_appliance"
-
-                # snapshot_yaml and num_tabs returned by generator
-                snapshot_yaml["panel_config"]["serial_number"] = f"{safe_device}_simulation" + (
-                    "" if suffix_index == 1 else f"_{suffix_index}"
-                )
-
-                # Ensure correct filename reflects computed num_tabs
-                base_name = f"simulation_config_{safe_device}_{num_tabs}_circuit.yaml"
-                dest_path = config_dir / base_name
-
-                # Ensure directory exists and write file
-                await self.hass.async_add_executor_job(
-                    lambda: dest_path.parent.mkdir(parents=True, exist_ok=True)
-                )
-
-                def _write_yaml() -> None:
-                    with dest_path.open("w", encoding="utf-8") as f:
-                        yaml.safe_dump(snapshot_yaml, f, sort_keys=False)
-
-                await self.hass.async_add_executor_job(_write_yaml)
-                _LOGGER.info("Cloned live panel to simulation YAML at %s", dest_path)
-
-                # Update config entry to point to the new simulation config
-                try:
-                    new_data = dict(self.config_entry.data)
-                    new_data[CONF_SIMULATION_CONFIG] = dest_path.stem
-                    self.hass.config_entries.async_update_entry(self.config_entry, data=new_data)
-                    _LOGGER.debug("Set CONF_SIMULATION_CONFIG to %s", dest_path.stem)
-                except Exception as update_err:
-                    _LOGGER.warning(
-                        "Failed to set CONF_SIMULATION_CONFIG to %s: %s",
-                        dest_path.stem,
-                        update_err,
-                    )
-
-                return self.async_create_entry(
-                    title="",
-                    data={},
-                    description=f"Cloned panel to {dest_path.name} in simulation_configs",
-                )
-
-            except Exception as e:
-                _LOGGER.error("Clone to simulation failed: %s", e)
-                errors["base"] = f"Clone failed: {e}"
 
         # Confirm form with destination field
         schema = vol.Schema(
@@ -1689,7 +1591,6 @@ class OptionsFlowHandler(config_entries.OptionsFlow):
             data_schema=schema,
             description_placeholders={
                 "panel": device_name or "Span Panel",
-                "tabs": str(num_tabs),
             },
             errors=errors,
         )
